app/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
blinker==1.7.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-rq==2.8.1
djangorestframework==3.14.0
Faker==20.0.3
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.18.3
MarkupSafe==2.1.3
msgpack==1.0.7
packaging==23.2
psutil==5.9.6
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
redis==5.0.1
requests==2.31.0
roundrobin==0.0.4
rq==1.15.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.1
